moving default keymap to dvorak
Initial implemenation of layers and keymap

Linux doesn't need a GUI Key

Getting rid of the gui keys on the default keymap.

adding tapdance for h/esc

swapping alt and ctr on mac

updating toggle

moving mac layer up in the stack

It seems like layer toggles like to move from top to bottom of the stack so I'm trying to move the mac layer to be more prominent.

Replacing caps with shift

The caps key isn't very useful with my modifiers on shift already
I'm moving it to be a plain shift key since it makes more sense in intellij

Updating with a jetbrains hiding macro

Adding jetbrains debug support to command layer